Added new profile Added new property to pom Added basic tests

+++ src/main/java/ch/erni/community/footsign/util/FileDownloader.java
@@ -24,7 +24,7 @@ public class FileDownloader {
 
 	private static final String PHOTO_URL = "https://sps2010-secure.erninet.ch/people/portraits/";
 	private static final String PHOTO_SUFFIX = "_197x245px_RGB.jpg";
-	private static final String MAVEN_DIR_PROPERTY = "photo.dir";
+	private static final String DEFAULT_PHOTO_NAME = "default_profile_photo.png";
 
 	@Autowired
 	PropertyLoader propertyLoader;
@@ -32,18 +32,20 @@ public class FileDownloader {
 	public Path downloadPhoto(UserDetails userDetails, String password) {
 
 		try {
+			File avatarsDir = new File(buildAvatarsPath());
+			if (!avatarsDir.exists()) {
+				return new File(buildRelativePath(null)).toPath();
+			}
 
+			Path target = new File(buildAvatarsPath() + buildPhotoName(userDetails)).toPath();
 			URL server = new URL(buildURLPhotoPath(userDetails));
-			Path target = new File(propertyLoader.getProperty(MAVEN_DIR_PROPERTY) + buildPhotoName(userDetails)).toPath();
-
 			Authenticator.setDefault(new SharePointAuthenticator(userDetails.getDomainUserName(), password));
 			Files.copy(server.openStream(), target, StandardCopyOption.REPLACE_EXISTING);
 
-			return target;
+			return new File(buildRelativePath(userDetails)).toPath();
 		} catch (PropertyFileNotFound | IOException e) {
 			throw new RuntimeException(e);
 		}
-
 	}
 
 	private String buildURLPhotoPath(UserDetails userDetails) {
@@ -51,7 +53,19 @@ public class FileDownloader {
 	}
 
 	private String buildPhotoName(UserDetails userDetails) {
+		if (userDetails == null) {
+			return DEFAULT_PHOTO_NAME;
+		}
 		return normalize(userDetails.getSecondName()) + "_" + normalize(userDetails.getFirstName()) + PHOTO_SUFFIX;
+	}
+
+	public String buildAvatarsPath() throws PropertyFileNotFound {
+		return propertyLoader.getProperty(PropertyConstatns.MAVEN_PROJECT_DIR_PROPERTY) +
+				propertyLoader.getProperty(PropertyConstatns.MAVEN_PHOTO_DIR_PROPERTY);
+	}
+
+	private String buildRelativePath(UserDetails userDetails) throws PropertyFileNotFound {
+		return propertyLoader.getProperty(PropertyConstatns.MAVEN_PHOTO_DIR_PROPERTY) + buildPhotoName(userDetails);
 	}
 
 	private String normalize(String input) {
